flag pinned connections invalid on network error
